pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask-cors>=6.0.1",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pillow>=11.3.0",
    "tensorflow>=2.20.0",
    "numpy>=2.3.2",
    "uuid>=1.30",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
    "bcrypt>=4.3.0",
    "qrcode>=8.2",
    "python-dotenv>=1.1.1",
]
